refactoring - file structure changes - auth and user management seperation
