+++ backend/src/main/java/com/interview/controller/UserController.java
@@ -1,13 +1,11 @@
 package com.interview.controller;
 
-import com.interview.dto.AdminUserRequest;
 import com.interview.dto.UserRequest;
 import com.interview.dto.UserResponse;
 import com.interview.service.UserService;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Positive;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
